initial modifications for nodejs and browserify
