+++ .gitignore
@@ -1,0 +1,4 @@
+node_modules/
+package-lock.json
+*bundle.js
+*.log
