+++ filters/ansi.js
@@ -1,3 +1,9 @@
+module.exports = {
+  ansi2html: ansi2html,
+  _get_extended_color: _get_extended_color,
+  _make_ansi_tags: _make_ansi_tags
+};
+
 _ANSI_COLORS = [
     'ansi-black',
     'ansi-red',
@@ -220,24 +226,4 @@ function _make_ansi_tags(fg, bg, bold, underline, inverse) {
   starttag += '>';
   
   return [starttag, '</span>']
-}
-
-
-if (
-    typeof module !== 'undefined' &&
-    module.exports &&
-    typeof exports !== 'undefined'
-  ) {
-  module.exports = {
-    ansi2html: ansi2html,
-    _get_extended_color: _get_extended_color,
-    _make_ansi_tags: _make_ansi_tags
-  };
-}
-else {
-  ansi = {
-    ansi2html: ansi2html,
-    _get_extended_color: _get_extended_color,
-    _make_ansi_tags: _make_ansi_tags
-  };
 }
